move functions in stat/utils.py to appropriate utils/* files
